Move modules -> exclude from coverage
