+++ app/src/androidTest/java/com/sdp13epfl2021/projmag/PreferencesActivityTests.kt
@@ -8,7 +8,7 @@ import androidx.test.ext.junit.rules.ActivityScenarioRule
 import com.sdp13epfl2021.projmag.JavaToKotlinHelperAndroidTest.anyObject
 import com.sdp13epfl2021.projmag.activities.PreferencesActivity
 import com.sdp13epfl2021.projmag.database.interfaces.UserdataDatabase
-import com.sdp13epfl2021.projmag.database.interfaces.UserdataDatabaseModule
+import com.sdp13epfl2021.projmag.database.di.UserdataDatabaseModule
 import com.sdp13epfl2021.projmag.model.ProjectFilter
 import dagger.hilt.android.testing.BindValue
 import dagger.hilt.android.testing.HiltAndroidRule
