+++ app/src/main/java/com/sdp13epfl2021/projmag/database/di/DatabaseModules.kt
@@ -1,11 +1,9 @@
-package com.sdp13epfl2021.projmag.database.interfaces
+package com.sdp13epfl2021.projmag.database.di
 
 import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.auth.ktx.auth
 import com.google.firebase.firestore.FirebaseFirestore
-import com.google.firebase.firestore.ktx.firestore
-import com.google.firebase.ktx.Firebase
 import com.sdp13epfl2021.projmag.database.impl.firebase.FirebaseUserdataDatabase
+import com.sdp13epfl2021.projmag.database.interfaces.UserdataDatabase
 import dagger.Binds
 import dagger.Module
 import dagger.Provides
